Fix win build
BUG=
R=hpayer@chromium.org

Review URL: https://chromiumcodereview.appspot.com/15934019

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14939 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/hydrogen-instructions.cc
@@ -2341,7 +2341,8 @@ Range* HBitwise::InferRange(Zone* zone) {
       if (right_lower < 0) right_lower = ~right_lower;
 
       int high = MostSignificantBit(
-          left_upper | left_lower | right_upper | right_lower);
+          static_cast<uint32_t>(
+              left_upper | left_lower | right_upper | right_lower));
 
       int64_t limit = 1;
       limit <<= high;
